+++ src/main/java/com/account_catalogue/taxes/infraestructure/adapters/output/jpaAdapters/entity/TaxEntity.java
@@ -35,7 +35,7 @@ public class TaxEntity {
     
     @Positive
     @Column(nullable = false)
-    private Float interest;
+    private Double interest;
 
     @ManyToOne
     @JoinColumn(name="depositAccount_code")
